refactor: reorganize web app components and extract shared utilities
