+++ apps/web/src/components/swimlane/stack-column.tsx
@@ -5,6 +5,7 @@ import type { BranchResponse, StackDetail } from "@/lib/api";
 import { BranchCard } from "@/components/branch-card";
 import { FolderGit2, ChevronDown } from "lucide-react";
 import Markdown from "react-markdown";
+import { stackStatusStyles } from "@/lib/status-config";
 
 export { shortenBranchName } from "@/components/branch-card";
 
@@ -94,37 +95,6 @@ export function StackColumn({
   );
 }
 
-const statusConfig: Record<string, { label: string; bg: string; selectedBg: string; text: string; shadow: string }> = {
-  shippable: {
-    label: "Ready to ship",
-    bg: "bg-green-100 dark:bg-green-950",
-    selectedBg: "bg-green-200 dark:bg-green-900",
-    text: "text-green-700 dark:text-green-400",
-    shadow: "shadow-[0_2px_8px_rgba(34,197,94,0.2)] dark:shadow-[0_2px_8px_rgba(34,197,94,0.15)]",
-  },
-  pending: {
-    label: "Needs restack",
-    bg: "bg-amber-100 dark:bg-amber-950",
-    selectedBg: "bg-amber-200 dark:bg-amber-900",
-    text: "text-amber-700 dark:text-amber-400",
-    shadow: "shadow-[0_2px_8px_rgba(245,158,11,0.2)] dark:shadow-[0_2px_8px_rgba(245,158,11,0.15)]",
-  },
-  blocked: {
-    label: "Blocked",
-    bg: "bg-red-100 dark:bg-red-950",
-    selectedBg: "bg-red-200 dark:bg-red-900",
-    text: "text-red-700 dark:text-red-400",
-    shadow: "shadow-[0_2px_8px_rgba(239,68,68,0.2)] dark:shadow-[0_2px_8px_rgba(239,68,68,0.15)]",
-  },
-  incomplete: {
-    label: "Incomplete",
-    bg: "bg-muted",
-    selectedBg: "bg-muted",
-    text: "text-muted-foreground",
-    shadow: "shadow-[0_2px_8px_rgba(0,0,0,0.05)] dark:shadow-[0_2px_8px_rgba(0,0,0,0.15)]",
-  },
-};
-
 export function StackStatusFooter({
   status,
   selected,
@@ -136,7 +106,7 @@ export function StackStatusFooter({
   onClick: () => void;
   compact?: boolean;
 }) {
-  const c = statusConfig[status] || statusConfig.incomplete;
+  const c = stackStatusStyles[status] || stackStatusStyles.incomplete;
 
   return (
     <button
