+++ apps/web/src/lib/github.ts
@@ -1,0 +1,7 @@
+export function prUrl(owner: string, repo: string, number: number): string {
+  return `https://github.com/${owner}/${repo}/pull/${number}`;
+}
+
+export function commitUrl(owner: string, repo: string, sha: string): string {
+  return `https://github.com/${owner}/${repo}/commit/${sha}`;
+}
